+++ src/render/mod.rs
@@ -1,5 +1,705 @@
 use crate::*;
 
+#[derive(Copy, Clone)]
+pub(super) struct Canvas {
+    ideal_num_xsteps: u32,
+    ideal_num_ysteps: u32,
+    width: f64,
+    height: f64,
+    padding: f64,
+    paddingy: f64,
+    aspect_offset: f64,
+    scalex2: f64,
+    scaley2: f64,
+    spacing: f64,
+    legendx1: f64,
+    num_css_classes: Option<usize>,
+    preserve_aspect: bool,
+}
+impl Canvas {
+    pub fn new() -> Self {
+        Self::with_options(false, None)
+    }
+    pub fn with_options(preserve_aspect: bool, num_css_classes: Option<usize>) -> Self {
+        let ideal_num_xsteps = if preserve_aspect { 4 } else { 6 };
+
+        let ideal_num_ysteps = 5;
+
+        let width = crate::WIDTH as f64;
+        let height = crate::HEIGHT as f64;
+        let padding = 150.0;
+        let paddingy = 100.0;
+
+        let aspect_offset = if preserve_aspect {
+            width / 2.0 - height + paddingy * 2.0
+        } else {
+            0.0
+        };
+
+        //The range over which the data will be scaled to fit
+        let scalex2 = if preserve_aspect {
+            height - paddingy * 2.0
+        } else {
+            width - padding * 2.0
+        };
+
+        let scaley2 = height - paddingy * 2.0;
+
+        let spacing = padding / 3.0;
+        let legendx1 = width - padding / 1.2 + padding / 30.0;
+
+        Canvas {
+            ideal_num_xsteps,
+            ideal_num_ysteps,
+            width,
+            height,
+            padding,
+            paddingy,
+            aspect_offset,
+            scalex2,
+            scaley2,
+            spacing,
+            legendx1,
+            num_css_classes,
+            preserve_aspect,
+        }
+    }
+
+    pub fn gen_ticks<X: PlotNum, Y: PlotNum>(
+        &self,
+        plotter: &PlotterRes<X, Y>,
+    ) -> TickResult<X, Y> {
+        let tickx = X::compute_ticks(
+            self.ideal_num_xsteps,
+            plotter.boundx,
+            DashInfo {
+                ideal_dash_size: 30.0,
+                max: self.scalex2,
+            },
+        );
+
+        let ticky = Y::compute_ticks(
+            self.ideal_num_ysteps,
+            plotter.boundy,
+            DashInfo {
+                ideal_dash_size: 30.0,
+                max: self.scaley2,
+            },
+        );
+
+        TickResult { tickx, ticky }
+    }
+
+    pub fn render<X: PlotNum, Y: PlotNum>(
+        &self,
+        writer: impl std::fmt::Write,
+        mut plotter: PlotterRes<X, Y>,
+        ticks: TickResult<X, Y>,
+    ) -> std::fmt::Result {
+        let Canvas {
+            width,
+            height,
+            padding,
+            paddingy,
+            aspect_offset,
+            scalex2,
+            scaley2,
+            spacing,
+            legendx1,
+            num_css_classes,
+            ..
+        } = *self;
+
+        let [minx, maxx] = plotter.boundx;
+        let [miny, maxy] = plotter.boundy;
+
+        let mut writer = tagger::new(writer);
+
+        let mut color_iter = {
+            let max = if let Some(nn) = num_css_classes {
+                nn
+            } else {
+                usize::MAX
+            };
+
+            (0..max).cycle()
+        };
+
+        for (i, mut p) in plotter.plots.drain(..).enumerate() {
+            let legendy1 = paddingy - padding / 8.0 + (i as f64) * spacing;
+
+            let name_exists = writer
+                .elem("text", |d| {
+                    d.attr("class", "poloto_text poloto_legend_text")?;
+                    d.attr("alignment-baseline", "middle")?;
+                    d.attr("text-anchor", "start")?;
+                    d.attr("font-size", "large")?;
+                    d.attr("x", width - padding / 1.2)?;
+                    d.attr("y", paddingy + (i as f64) * spacing)
+                })?
+                .build(|d| {
+                    let mut wc = num::WriteCounter::new(d.writer_safe());
+                    p.plots.write_name(&mut wc)?;
+                    Ok(wc.get_counter() != 0)
+                })?;
+
+            let aa = minx.scale([minx, maxx], scalex2);
+            let bb = miny.scale([miny, maxy], scaley2);
+
+            struct PlotIter<X, Y> {
+                basex_ii: f64,
+                basey_ii: f64,
+                rangex_ii: [X; 2],
+                rangey_ii: [Y; 2],
+                maxx_ii: f64,
+                maxy_ii: f64,
+            }
+            impl<X: PlotNum, Y: PlotNum> PlotIter<X, Y> {
+                fn gen_iter<'a>(
+                    &'a self,
+                    p: &'a mut Plot<X, Y>,
+                ) -> impl Iterator<Item = [f64; 2]> + 'a {
+                    p.plots.iter_second().map(move |(x, y)| {
+                        [
+                            self.basex_ii + x.scale(self.rangex_ii, self.maxx_ii),
+                            self.basey_ii - y.scale(self.rangey_ii, self.maxy_ii),
+                        ]
+                    })
+                }
+            }
+
+            let plot_iter = PlotIter {
+                basex_ii: aspect_offset + padding - aa,
+                basey_ii: height - paddingy + bb,
+                rangex_ii: [minx, maxx],
+                rangey_ii: [miny, maxy],
+                maxx_ii: scalex2,
+                maxy_ii: scaley2,
+            };
+
+            let colori = color_iter.next().unwrap();
+
+            match p.plot_type {
+                PlotType::Line => {
+                    if name_exists {
+                        writer.single("line", |d| {
+                            d.attr(
+                                "class",
+                                format_args!(
+                                    "poloto_line poloto_legend_icon poloto{}stroke poloto{}legend",
+                                    colori, colori
+                                ),
+                            )?;
+                            d.attr("stroke", "black")?;
+                            d.attr("x1", legendx1)?;
+                            d.attr("x2", legendx1 + padding / 3.0)?;
+                            d.attr("y1", legendy1)?;
+                            d.attr("y2", legendy1)
+                        })?;
+                    }
+
+                    writer.single("path", |d| {
+                        d.attr("class", format_args!("poloto_line poloto{}stroke", colori))?;
+                        d.attr("fill", "none")?;
+                        d.attr("stroke", "black")?;
+                        d.path(|a| render::line(a, plot_iter.gen_iter(&mut p)))
+                    })?;
+                }
+                PlotType::Scatter => {
+                    if name_exists {
+                        writer.single("line", |d| {
+                            d.attr(
+                                "class",
+                                format_args!(
+                                    "poloto_scatter poloto_legend_icon poloto{}stroke poloto{}legend",
+                                    colori, colori
+                                ),
+                            )?;
+                            d.attr("stroke", "black")?;
+                            d.attr("x1", legendx1 + padding / 30.0)?;
+                            d.attr("x2", legendx1 + padding / 30.0)?;
+                            d.attr("y1", legendy1)?;
+                            d.attr("y2", legendy1)
+                        })?;
+                    }
+
+                    writer.single("path", |d| {
+                        d.attr(
+                            "class",
+                            format_args!("poloto_scatter poloto{}stroke", colori),
+                        )?;
+                        d.path(|a| {
+                            use tagger::PathCommand::*;
+                            for [x, y] in plot_iter
+                                .gen_iter(&mut p)
+                                .filter(|&[x, y]| x.is_finite() && y.is_finite())
+                            {
+                                a.put(M(x, y))?;
+                                a.put(H_(0))?;
+                            }
+                            Ok(())
+                        })
+                    })?;
+                }
+                PlotType::Histo => {
+                    if name_exists {
+                        writer.single("rect", |d| {
+                            d.attr(
+                                "class",
+                                format_args!(
+                                    "poloto_histo poloto_legend_icon poloto{}fill poloto{}legend",
+                                    colori, colori
+                                ),
+                            )?;
+                            d.attr("x", legendx1)?;
+                            d.attr("y", legendy1 - padding / 30.0)?;
+                            d.attr("width", padding / 3.0)?;
+                            d.attr("height", padding / 20.0)?;
+                            d.attr("rx", padding / 30.0)?;
+                            d.attr("ry", padding / 30.0)
+                        })?;
+                    }
+
+                    writer
+                        .elem("g", |d| {
+                            d.attr("class", format_args!("poloto_histo poloto{}fill", colori))
+                        })?
+                        .build(|writer| {
+                            let mut last = None;
+                            //TODO dont necesarily filter?
+                            for [x, y] in plot_iter
+                                .gen_iter(&mut p)
+                                .filter(|&[x, y]| x.is_finite() && y.is_finite())
+                            {
+                                if let Some((lx, ly)) = last {
+                                    writer.single("rect", |d| {
+                                        d.attr("x", lx)?;
+                                        d.attr("y", ly)?;
+                                        d.attr(
+                                            "width",
+                                            (padding * 0.02).max((x - lx) - (padding * 0.02)),
+                                        )?;
+                                        d.attr("height", height - paddingy - ly)
+                                    })?;
+                                }
+                                last = Some((x, y))
+                            }
+                            Ok(())
+                        })?;
+                }
+                PlotType::LineFill => {
+                    if name_exists {
+                        writer.single("rect", |d| {
+                            d.attr(
+                                "class",
+                                format_args!(
+                                    "poloto_linefill poloto_legend_icon poloto{}fill poloto{}legend",
+                                    colori, colori
+                                ),
+                            )?;
+                            d.attr("x", legendx1)?;
+                            d.attr("y", legendy1 - padding / 30.0)?;
+                            d.attr("width", padding / 3.0)?;
+                            d.attr("height", padding / 20.0)?;
+                            d.attr("rx", padding / 30.0)?;
+                            d.attr("ry", padding / 30.0)
+                        })?;
+                    }
+
+                    writer.single("path", |d| {
+                        d.attr(
+                            "class",
+                            format_args!("poloto_linefill poloto{}fill", colori),
+                        )?;
+                        d.path(|path| {
+                            render::line_fill(
+                                path,
+                                plot_iter.gen_iter(&mut p),
+                                height - paddingy,
+                                true,
+                            )
+                        })
+                    })?;
+                }
+                PlotType::LineFillRaw => {
+                    if name_exists {
+                        writer.single("rect", |d| {
+                            d.attr(
+                                "class",
+                                format_args!(
+                                    "poloto_linefillraw poloto_legend_icon poloto{}fill poloto{}legend",
+                                    colori, colori
+                                ),
+                            )?;
+                            d.attr("x", legendx1)?;
+                            d.attr("y", legendy1 - padding / 30.0)?;
+                            d.attr("width", padding / 3.0)?;
+                            d.attr("height", padding / 20.0)?;
+                            d.attr("rx", padding / 30.0)?;
+                            d.attr("ry", padding / 30.0)
+                        })?;
+                    }
+
+                    writer.single("path", |d| {
+                        d.attr(
+                            "class",
+                            format_args!("poloto_linefillraw poloto{}fill", colori),
+                        )?;
+                        d.path(|path| {
+                            render::line_fill(
+                                path,
+                                plot_iter.gen_iter(&mut p),
+                                height - paddingy,
+                                false,
+                            )
+                        })
+                    })?;
+                }
+            }
+        }
+
+        self.draw_base(&mut writer, plotter, ticks)?;
+
+        Ok(())
+    }
+
+    pub fn draw_base<X: PlotNum, Y: PlotNum, T: fmt::Write>(
+        &self,
+        writer: &mut tagger::ElemWriter<T>,
+        mut plotter: PlotterRes<X, Y>,
+        ticks: TickResult<X, Y>,
+    ) -> std::fmt::Result {
+        let Canvas {
+            width,
+            height,
+            padding,
+            paddingy,
+            aspect_offset,
+            scalex2,
+            scaley2,
+            preserve_aspect,
+            ..
+        } = *self;
+
+        let boundx = plotter.boundx;
+        let boundy = plotter.boundy;
+        let [minx, maxx] = boundx;
+        let [miny, maxy] = boundy;
+
+        let xtick_info = ticks.tickx;
+        let ytick_info = ticks.ticky;
+
+        let texty_padding = paddingy * 0.3;
+        let textx_padding = padding * 0.1;
+
+        writer
+            .elem("text", |d| {
+                d.attr("class", "poloto_labels poloto_text poloto_title")?;
+                d.attr("alignment-baseline", "start")?;
+                d.attr("text-anchor", "middle")?;
+                d.attr("font-size", "x-large")?;
+                d.attr("x", width / 2.0)?;
+                d.attr("y", padding / 4.0)
+            })?
+            .build(|w| {
+                plotter.title.fmt_self(Data {
+                    writer: &mut w.writer_safe(),
+                    boundx,
+                    boundy,
+                    stepx: xtick_info.unit_data,
+                    stepy: ytick_info.unit_data,
+                })
+            })?;
+
+        writer
+            .elem("text", |d| {
+                d.attr("class", "poloto_labels poloto_text poloto_xname")?;
+                d.attr("alignment-baseline", "start")?;
+                d.attr("text-anchor", "middle")?;
+                d.attr("font-size", "x-large")?;
+                d.attr("x", width / 2.0)?;
+                d.attr("y", height - padding / 8.)
+            })?
+            .build(|w| {
+                plotter.xname.fmt_self(Data {
+                    writer: &mut w.writer_safe(),
+                    boundx,
+                    boundy,
+                    stepx: xtick_info.unit_data,
+                    stepy: ytick_info.unit_data,
+                })
+            })?;
+
+        writer
+            .elem("text", |d| {
+                d.attr("class", "poloto_labels poloto_text poloto_yname")?;
+                d.attr("alignment-baseline", "start")?;
+                d.attr("text-anchor", "middle")?;
+                d.attr("font-size", "x-large")?;
+                d.attr(
+                    "transform",
+                    format_args!("rotate(-90,{},{})", padding / 4.0, height / 2.0),
+                )?;
+                d.attr("x", padding / 4.0)?;
+                d.attr("y", height / 2.0)
+            })?
+            .build(|w| {
+                plotter.yname.fmt_self(Data {
+                    writer: &mut w.writer_safe(),
+                    boundx,
+                    boundy,
+                    stepx: xtick_info.unit_data,
+                    stepy: ytick_info.unit_data,
+                })
+            })?;
+
+        let xdash_size: Option<f64> = if plotter.dash_x {
+            xtick_info.dash_size
+        } else {
+            None
+        };
+        let ydash_size: Option<f64> = if plotter.dash_y {
+            ytick_info.dash_size
+        } else {
+            None
+        };
+
+        use tagger::PathCommand::*;
+
+        let first_tickx = xtick_info.ticks[0];
+
+        let first_ticky = ytick_info.ticks[0];
+
+        {
+            //step num is assured to be atleast 1.
+            let extra = if let Some(base) = xtick_info.display_relative {
+                writer
+                    .elem("text", |d| {
+                        d.attr("class", "poloto_tick_labels poloto_text")?;
+                        d.attr("alignment-baseline", "middle")?;
+                        d.attr("text-anchor", "start")?;
+                        d.attr("x", width * 0.55)?;
+                        d.attr("y", paddingy * 0.7)
+                    })?
+                    .build(|d| {
+                        let mut w = d.writer_safe();
+                        use std::fmt::Write;
+                        write!(w, "Where j = ")?;
+
+                        plotter.xtick_fmt.fmt_self(
+                            base,
+                            DataSingle {
+                                writer: &mut w,
+                                bound: boundx,
+                                ff: FmtFull::Full,
+                                step: xtick_info.unit_data,
+                            },
+                        )
+                    })?;
+
+                "j+"
+            } else {
+                ""
+            };
+
+            //Draw interva`l x text
+            for &Tick { position, value } in xtick_info.ticks.iter() {
+                let xx = (position.scale([minx, maxx], scalex2)
+                    - minx.scale([minx, maxx], scalex2))
+                    + padding;
+
+                writer.single("line", |d| {
+                    d.attr("class", "poloto_axis_lines")?;
+                    d.attr("stroke", "black")?;
+                    d.attr("x1", aspect_offset + xx)?;
+                    d.attr("x2", aspect_offset + xx)?;
+                    d.attr("y1", height - paddingy)?;
+                    d.attr("y2", height - paddingy * 0.95)
+                })?;
+
+                writer
+                    .elem("text", |d| {
+                        d.attr("class", "poloto_tick_labels poloto_text")?;
+                        d.attr("alignment-baseline", "start")?;
+                        d.attr("text-anchor", "middle")?;
+                        d.attr("x", aspect_offset + xx)?;
+                        d.attr("y", height - paddingy + texty_padding)
+                    })?
+                    .build(|w| {
+                        let mut w = w.writer_safe();
+                        use std::fmt::Write;
+                        write!(w, "{}", extra)?;
+
+                        plotter.xtick_fmt.fmt_self(
+                            value,
+                            DataSingle {
+                                writer: &mut w,
+                                bound: boundx,
+                                ff: FmtFull::Short,
+                                step: xtick_info.unit_data,
+                            },
+                        )
+                        /*
+                        w.put_raw(format_args!(
+                            "{}{}",
+                            extra,
+                            DisplayableClosure::new(|w| plotter.xcontext.fmt_tick(
+                                w,
+                                value,
+                                xtick_info.unit_data,
+                                FmtFull::Tick
+                            ))
+                        ))
+                        */
+                    })?;
+            }
+        }
+
+        {
+            //step num is assured to be atleast 1.
+            let extra = if let Some(base) = ytick_info.display_relative {
+                writer
+                    .elem("text", |d| {
+                        d.attr("class", "poloto_tick_labels poloto_text")?;
+                        d.attr("alignment-baseline", "middle")?;
+                        d.attr("text-anchor", "start")?;
+                        d.attr("x", padding)?;
+                        d.attr("y", paddingy * 0.7)
+                    })?
+                    .build(|w| {
+                        use std::fmt::Write;
+                        let mut w = w.writer_safe();
+                        write!(w, "Where k = ")?;
+
+                        plotter.ytick_fmt.fmt_self(
+                            base,
+                            DataSingle {
+                                writer: &mut w,
+                                bound: boundy,
+                                ff: FmtFull::Full,
+                                step: ytick_info.unit_data,
+                            },
+                        )
+                    })?;
+
+                "k+"
+            } else {
+                ""
+            };
+
+            //Draw interval y text
+            for &Tick { position, value } in ytick_info.ticks.iter() {
+                let yy = height
+                    - (position.scale([miny, maxy], scaley2) - miny.scale([miny, maxy], scaley2))
+                    - paddingy;
+
+                writer.single("line", |d| {
+                    d.attr("class", "poloto_axis_lines")?;
+                    d.attr("stroke", "black")?;
+                    d.attr("x1", aspect_offset + padding)?;
+                    d.attr("x2", aspect_offset + padding * 0.96)?;
+                    d.attr("y1", yy)?;
+                    d.attr("y2", yy)
+                })?;
+
+                writer
+                    .elem("text", |d| {
+                        d.attr("class", "poloto_tick_labels poloto_text")?;
+                        d.attr("alignment-baseline", "middle")?;
+                        d.attr("text-anchor", "end")?;
+                        d.attr("x", aspect_offset + padding - textx_padding)?;
+                        d.attr("y", yy)
+                    })?
+                    .build(|w| {
+                        let mut w = w.writer_safe();
+                        use std::fmt::Write;
+                        write!(w, "{}", extra)?;
+
+                        plotter.ytick_fmt.fmt_self(
+                            value,
+                            DataSingle {
+                                writer: &mut w,
+                                bound: boundy,
+                                ff: FmtFull::Short,
+                                step: ytick_info.unit_data,
+                            },
+                        )
+
+                        /*
+                        w.put_raw(format_args!(
+                            "{}{}",
+                            extra,
+                            DisplayableClosure::new(|w| plotter.ycontext.fmt_tick(
+                                w,
+                                value,
+                                ytick_info.unit_data,
+                                FmtFull::Tick
+                            )) //TODO need a way to communicate writing base
+                        ))
+                        */
+                    })?;
+            }
+        }
+
+        let d1 = minx.scale([minx, maxx], scalex2);
+        let d2 = first_tickx.position.scale([minx, maxx], scalex2);
+        let distance_to_firstx = d2 - d1;
+
+        writer.single("path", |d| {
+            d.attr("stroke", "black")?;
+            d.attr("fill", "none")?;
+            d.attr("class", "poloto_axis_lines")?;
+            if let Some(xdash_size) = xdash_size {
+                d.attr(
+                    "style",
+                    format_args!(
+                        "stroke-dasharray:{};stroke-dashoffset:{};",
+                        xdash_size / 2.0,
+                        -distance_to_firstx
+                    ),
+                )?;
+            }
+            d.path(|p| {
+                p.put(M(padding + aspect_offset, height - paddingy))?;
+                if preserve_aspect {
+                    p.put(L(
+                        height - paddingy / 2.0 + aspect_offset,
+                        height - paddingy,
+                    ))
+                } else {
+                    p.put(L(width - padding + aspect_offset, height - paddingy))
+                }
+            })
+        })?;
+
+        let d1 = miny.scale([miny, maxy], scaley2);
+        let d2 = first_ticky.position.scale([miny, maxy], scaley2);
+        let distance_to_firsty = d2 - d1;
+
+        writer.single("path", |d| {
+            d.attr("stroke", "black")?;
+            d.attr("fill", "none")?;
+            d.attr("class", "poloto_axis_lines")?;
+            if let Some(ydash_size) = ydash_size {
+                d.attr(
+                    "style",
+                    format_args!(
+                        "stroke-dasharray:{};stroke-dashoffset:{};",
+                        ydash_size / 2.0,
+                        -distance_to_firsty
+                    ),
+                )?;
+            }
+            d.path(|p| {
+                p.put(M(aspect_offset + padding, height - paddingy))?;
+                p.put(L(aspect_offset + padding, paddingy))
+            })
+        })?;
+
+        Ok(())
+    }
+}
+
 pub fn line_fill<T: std::fmt::Write>(
     path: &mut tagger::PathBuilder<T>,
     mut it: impl Iterator<Item = [f64; 2]>,
